Fix MSVC warnings;
It's that time of year.

+++ src/api/api.c
@@ -171,7 +171,7 @@ int _luax_checkenum(lua_State* L, int index, const StringEntry* map, const char*
   size_t length;
   const char* string = fallback ? luaL_optlstring(L, index, fallback, &length) : luaL_checklstring(L, index, &length);
 
-  for (size_t i = 0; map[i].length; i++) {
+  for (int i = 0; map[i].length; i++) {
     if (map[i].length == length && !memcmp(map[i].string, string, length)) {
       return i;
     }
